--- docs/source/python_api/mlflow.models.signature.rst
@@ -1,6 +1,0 @@
-mlflow.models.signature
-=======================
-.. automodule:: mlflow.models.signature
-    :members:
-    :undoc-members:
-    :show-inheritance:
