--- docs/source/python_api/mlflow.models.utils.rst
@@ -1,6 +1,0 @@
-mlflow.models.utils
-===================
-.. automodule:: mlflow.models.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
