+++ docs/source/python_api/mlflow.types.rst
@@ -1,6 +1,8 @@
 mlflow.types
-============
+==============
+
 .. automodule:: mlflow.types
     :members:
     :undoc-members:
     :show-inheritance:
+
