--- docs/source/python_api/mlflow.types.schema.rst
@@ -1,6 +1,0 @@
-mlflow.types.schema
-===================
-.. automodule:: mlflow.types.schema
-    :members:
-    :undoc-members:
-    :show-inheritance:
