+++ mlflow/h2o.py
@@ -135,23 +135,28 @@ def log_model(h2o_model, artifact_path, conda_env=None, registered_model_name=No
                                   version under ``registered_model_name``, also creating a
                                   registered model if one with the given name does not exist.
 
-   :param signature: Note:: Experimental: This argument may change or be removed in a
-                     future release without warning. Model signature describes model input
-                     and output schema. The model signature can be inferred from datasets
-                     representing valid model input (e.g. the training dataset) and valid
-                     model output (e.g. model predictions generated on the training dataset).
-                     For example, you can obtain model signature as follows:
+    :param signature: Note:: Experimental: This argument may change or be removed
+                      release without warning.
 
-                     .. code-block:: python
+                      :py:class:`ModelSignature <mlflow.models.ModelSignature>` describes model
+                      input and output :py:class:`Schema <mlflow.types.Schema>`. The model
+                      signature can be :py:func:`inferred <mlflow.models.infer_signature>` from
+                      datasets with valid model input (e.g. the training dataset) and valid
+                      model output (e.g. model predictions generated on the training dataset),
+                      for example:
 
-                       from mlflow.models.signature import infer_signature
-                       train = df.drop_column("target_label")
-                       signature = infer_signature(train, model.predict(train))
+                      .. code-block:: python
 
-    :param input_example: Note:: Experimental: This argument may change or be removed in a
-                          future release without warning. Input example provides one or several
-                          examples of valid model input. The example can be used as a hint of what
-                          data to feed the model.
+                        from mlflow.models.signature import infer_signature
+                        train = df.drop_column("target_label")
+                        signature = infer_signature(train, model.predict(train))
+    :param input_example: Note:: Experimental: This argument may change or be removed in a future
+                          release without warning. Input example provides one or several instances
+                          of valid model input. The example can be used as a hint of what data to
+                          feed the model. The given example will be converted to a Pandas DataFrame
+                          and then serialized to json using the Pandas split-oriented format. Bytes
+                          are base64-encoded.
+
     :param kwargs: kwargs to pass to ``h2o.save_model`` method.
     """
     Model.log(artifact_path=artifact_path, flavor=mlflow.h2o,
