+++ mlflow/spacy.py
@@ -19,7 +19,8 @@ import yaml
 import mlflow
 from mlflow import pyfunc
 from mlflow.exceptions import MlflowException
-from mlflow.models import Model
+from mlflow.models import Model, ModelSignature
+from mlflow.models.utils import ModelInputExample
 from mlflow.tracking.artifact_utils import _download_artifact_from_uri
 from mlflow.utils.environment import _mlflow_conda_env
 from mlflow.utils.model_utils import _get_flavor_configuration
@@ -111,7 +112,8 @@ def save_model(spacy_model, path, conda_env=None, mlflow_model=Model()):
     mlflow_model.save(os.path.join(path, "MLmodel"))
 
 
-def log_model(spacy_model, artifact_path, conda_env=None, registered_model_name=None, **kwargs):
+def log_model(spacy_model, artifact_path, conda_env=None, registered_model_name=None,
+              signature: ModelSignature = None, input_example: ModelInputExample = None, **kwargs):
     """
     Log a spaCy model as an MLflow artifact for the current run.
 
@@ -139,11 +141,34 @@ def log_model(spacy_model, artifact_path, conda_env=None, registered_model_name=
                                   future release without warning. If given, create a model
                                   version under ``registered_model_name``, also creating a
                                   registered model if one with the given name does not exist.
+    :param signature: Note:: Experimental: This argument may change or be removed
+                      release without warning.
+
+                      :py:class:`ModelSignature <mlflow.models.ModelSignature>` describes model
+                      input and output :py:class:`Schema <mlflow.types.Schema>`. The model
+                      signature can be :py:func:`inferred <mlflow.models.infer_signature>` from
+                      datasets with valid model input (e.g. the training dataset) and valid
+                      model output (e.g. model predictions generated on the training dataset),
+                      for example:
+
+                      .. code-block:: python
+
+                        from mlflow.models.signature import infer_signature
+                        train = df.drop_column("target_label")
+                        signature = infer_signature(train, model.predict(train))
+    :param input_example: Note:: Experimental: This argument may change or be removed in a future
+                          release without warning. Input example provides one or several instances
+                          of valid model input. The example can be used as a hint of what data to
+                          feed the model. The given example will be converted to a Pandas DataFrame
+                          and then serialized to json using the Pandas split-oriented format. Bytes
+                          are base64-encoded.
+
     :param kwargs: kwargs to pass to ``spacy.save_model`` method.
     """
     Model.log(artifact_path=artifact_path, flavor=mlflow.spacy,
               registered_model_name=registered_model_name,
-              spacy_model=spacy_model, conda_env=conda_env, **kwargs)
+              spacy_model=spacy_model, conda_env=conda_env,
+              signature=signature, input_example=input_example, **kwargs)
 
 
 def _load_model(path):
